Extract Store::{enqueue, enqueue_bulk} into impl module.
Following on from bbc4a6105861c4b5c4eff0cdbac54eabe502eb2b this time for
the enqueue operations. This is a pretty large extraction.
